# applications/storium-microservice/infra/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Storium API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/orders/checkout": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Checkout",
        "operationId": "checkout_api_orders_checkout_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Cart-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Cart-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOutputDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "List Orders",
        "operationId": "list_orders_api_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderOutputDTO"
                  },
                  "type": "array",
                  "title": "Response List Orders Api Orders Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/orders/{order_id}": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Order Detail",
        "operationId": "order_detail_api_orders__order_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOutputDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutInputDTO": {
        "properties": {
          "shipping_name": {
            "type": "string",
            "title": "Shipping Name"
          },
          "shipping_address": {
            "type": "string",
            "title": "Shipping Address"
          },
          "shipping_city": {
            "type": "string",
            "title": "Shipping City"
          },
          "shipping_phone": {
            "type": "string",
            "title": "Shipping Phone",
            "default": ""
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "shipping_name",
          "shipping_address",
          "shipping_city"
        ],
        "title": "CheckoutInputDTO"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OrderItemOutputDTO": {
        "properties": {
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "unit_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Unit Price"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "line_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Line Total"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "product_name",
          "unit_price",
          "quantity",
          "line_total"
        ],
        "title": "OrderItemOutputDTO"
      },
      "OrderOutputDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_display": {
            "type": "string",
            "title": "Status Display"
          },
          "total_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Price"
          },
          "shipping_name": {
            "type": "string",
            "title": "Shipping Name"
          },
          "shipping_address": {
            "type": "string",
            "title": "Shipping Address"
          },
          "shipping_city": {
            "type": "string",
            "title": "Shipping City"
          },
          "shipping_phone": {
            "type": "string",
            "title": "Shipping Phone"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemOutputDTO"
            },
            "type": "array",
            "title": "Items"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "status_display",
          "total_price",
          "shipping_name",
          "shipping_address",
          "shipping_city",
          "shipping_phone",
          "notes",
          "items",
          "created_at",
          "updated_at"
        ],
        "title": "OrderOutputDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}